Update Windows Tweaks.bat
This makes sure Recall is disabled even when the computer doesn't have the full version installed. On most modern Windows 11 computers, Recall is enabled but inactive, and the WindowsAI registry entry and group policy are not yet available. While Recall doesn't have an exe in these versions, there are still dlls and other files that exist, and Windows can be updated at any moment to roll out the full version of recall to more computers. The added line makes sure that Recall is disabled even in Windows installations where the full version does not exist yet.